Created HoloCell and shifted around some responsibilities.
