--- Arena.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 30c9b9482f2ce4448918e18cf83a1357
-timeCreated: 1489426491
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
